fix: rename plugin-data to state
