+++ designer-demo/registry.js
@@ -25,7 +25,7 @@ import {
   Refresh,
   Collaboration,
   Materials,
-  Data,
+  State,
   Script,
   Tree,
   Help,
@@ -77,7 +77,7 @@ export default {
     Fullscreen,
     Lang
   ],
-  plugins: [Materials, Tree, Page, Block, Datasource, Bridge, I18n, Script, Data, Schema, Help, Robot],
+  plugins: [Materials, Tree, Page, Block, Datasource, Bridge, I18n, Script, State, Schema, Help, Robot],
   dsls: [{ id: 'engine.dsls.dslvue' }],
   settings: [Props, Styles, Events],
   canvas: Canvas
