+++ jsconfig.json
@@ -8,7 +8,7 @@
       "@opentiny/tiny-engine-meta-register": ["packages/register/src/index.js"],
       "@opentiny/tiny-engine-canvas": ["packages/canvas/src/index.js"],
       "@opentiny/tiny-engine-plugin-materials": ["packages/plugins/materials/index"],
-      "@opentiny/tiny-engine-plugin-data": ["packages/plugins/data/index"],
+      "@opentiny/tiny-engine-plugin-state": ["packages/plugins/state/index"],
       "@opentiny/tiny-engine-plugin-script": ["packages/plugins/script/index"],
       "@opentiny/tiny-engine-plugin-tree": ["packages/plugins/tree/index"],
       "@opentiny/tiny-engine-plugin-help": ["packages/plugins/help/index"],
@@ -37,7 +37,7 @@
       "@opentiny/tiny-engine-svgs": ["packages/svgs/index"],
       "@opentiny/tiny-engine-http": ["packages/http/index"],
       "@opentiny/tiny-engine-plugin-materials/*": ["packages/plugins/materials/*"],
-      "@opentiny/tiny-engine-plugin-data/*": ["packages/plugins/data/*"],
+      "@opentiny/tiny-engine-plugin-state/*": ["packages/plugins/state/*"],
       "@opentiny/tiny-engine-plugin-script/*": ["packages/plugins/script/*"],
       "@opentiny/tiny-engine-plugin-tree/*": ["packages/plugins/tree/*"],
       "@opentiny/tiny-engine-plugin-help/*": ["packages/plugins/help/*"],
