+++ packages/design-core/index.js
@@ -16,7 +16,7 @@ export { default as Refresh } from '@opentiny/tiny-engine-toolbar-refresh'
 export { default as Collaboration } from '@opentiny/tiny-engine-toolbar-collaboration'
 export { default as Setting } from '@opentiny/tiny-engine-toolbar-setting'
 export { default as Materials, ResourceService, MaterialService } from '@opentiny/tiny-engine-plugin-materials'
-export { default as Data } from '@opentiny/tiny-engine-plugin-data'
+export { default as State } from '@opentiny/tiny-engine-plugin-state'
 export { default as Script } from '@opentiny/tiny-engine-plugin-script'
 export { default as Tree } from '@opentiny/tiny-engine-plugin-tree'
 export { default as Help, HelpService } from '@opentiny/tiny-engine-plugin-help'
