--- packages/plugins/data/src/js/constants.js
@@ -1,22 +1,0 @@
-/**
-* Copyright (c) 2023 - present TinyEngine Authors.
-* Copyright (c) 2023 - present Huawei Cloud Computing Technologies Co., Ltd.
-*
-* Use of this source code is governed by an MIT-style license.
-*
-* THE OPEN SOURCE SOFTWARE IN THIS PRODUCT IS DISTRIBUTED IN THE HOPE THAT IT WILL BE USEFUL,
-* BUT WITHOUT ANY WARRANTY, WITHOUT EVEN THE IMPLIED WARRANTY OF MERCHANTABILITY OR FITNESS FOR
-* A PARTICULAR PURPOSE. SEE THE APPLICABLE LICENSES FOR MORE DETAILS.
-*
-*/
-
-export const STATE = {
-  CURRENT_STATE: 'current',
-  GLOBAL_STATE: 'global'
-}
-
-export const OPTION_TYPE = {
-  ADD: 'add',
-  UPDATE: 'update',
-  COPY: 'copy'
-}
