+++ packages/plugins/state/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@opentiny/tiny-engine-plugin-data",
+  "name": "@opentiny/tiny-engine-plugin-state",
   "version": "1.0.3",
   "publishConfig": {
     "access": "public"
@@ -15,7 +15,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/opentiny/tiny-engine",
-    "directory": "packages/plugins/data"
+    "directory": "packages/plugins/state"
   },
   "bugs": {
     "url": "https://github.com/opentiny/tiny-engine/issues"
